[COOK-3642] Add Windows support
This pull request adds windows support to the ark package. It makes use
of the 7-zip cookbook to download and install 7-zip on windows to
extract the files.
As files are handled a little different on windows then on unix systems
a new parameter 'win_install_dir' is introduced. To this folder the
files are extracted to.
Unfortunately, there is no windows system in the kitchen config, so this
cannot be tested automatically, but tests were done.

Ticket: https://tickets.opscode.com/browse/COOK-3642