Rubocop Offenses
The simple attr issue in the spec_helper is a dumb one so I'm going
to block it on the spec_helper as my uses are different and I don't want
to have the presence of an equals sign.

The other things were small line issues spread throughout.

Refactored to use each_with_object to quiet down rubocop as well.