Refactored paths and command methods
This is a rudimentary refactoring to get all the helper methods that
I created before out of the helpers class and use a defaults class
to figure out all that information.

Also the specs are in a temporary state at the moment because of the
allow_any_instance_of ... which is seriously brute force.

I also added some TODO messages because some of the behavior seemed
strange and something that I want to explore later.