Merging #82 caused spec failures on Windows
It appears that when you 'suggest' a cookbook that does not load the cookbooks at all. So I added them as dependencies in the Berksfile. Which made the first error message disappear about the 7-zip is not a dependency. However, it seems that does not solve the final issue that the node attributes are not correctly loaded for the recipe that does get included through "include_recipe".