Notify "clean up" ruby block from target directory.
The "unpack" resource is notified from both the directory and
remote_file resources. Therefore, to ensure that "clean up" is executed
before unpack, "clean up" resource must also be notified from the
"directory" resource.

In the previous commit, the notification to the "clean up" resource was
only sent from the remote_file resource. Consequently, whenever the
target directory has to be created, a notification to the "unpack"
resource is first sent, which leads to unpack the archive then delete
the content of the target directory.