Support in-place upgrades using the cherry-pick action
Previously, a new version of the archive would not result in
a cherry-picked file to be extracted again because it already
exists. Removing the "creates" property allows in-place
upgrades and is still idempotent because the extraction command
only runs when a newer version of the archive is downloaded.

This makes cherry-pick logic similar to other actions' logic,
where no additional guards are used beyond chaining resources
with notifications.

Signed-off-by: Max Timchenko <max@maxvt.com>